refactor: 💡 separate video style from functionality
