+++ apps/e-learning-app/src/app/components/video/VideoStyle.js
@@ -1,0 +1,60 @@
+const videoStyle = {
+  video_style: {
+    fontFamily: 'Open Sans, sans-serif',
+    padding: '5px 20px 5px 0',
+    minWidth: '300px',
+    maxWidth: '300px',
+    minHeight: '300px',
+    maxHeight: '300px',
+  },
+  title_style: {
+    fontSize: '20px',
+    overflowWrap: 'break-word',
+    margin: '10px 0',
+  },
+  title_link_style: {
+    color: '#000',
+    textDecoration: 'none',
+  },
+  channel_style: {
+    fontSize: '16px',
+    marginBottom: '5px',
+    paddingRight: '0%',
+    textAlign: 'right',
+    overflowWrap: 'break-word',
+  },
+  channel_link_style: {
+    color: 'rgb(0,101,101)',
+    textDecoration: 'none',
+  },
+  thumbnail_style: {
+    position: 'relative',
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    overflow: 'hidden',
+  },
+  length_style: {
+    position: 'absolute',
+    background: 'black',
+    color: 'white',
+    fontSize: '12px',
+    padding: '5px 5px 5px 5px',
+    bottom: '3px',
+    left: '0px',
+  },
+
+  thumbnail_link_style: {
+    minWidth: '100%',
+    minHeight: '100%',
+    borderRadius: '2px',
+  },
+
+  img_style: {
+    minWidth: '100%',
+    minHeight: '100%',
+    borderRadius: '2px',
+  },
+};
+
+export default videoStyle;
